Compilation checkpoint

+++ src/main/java/se/oneagency/codechallenge/elevator/resources/OutsideRequest.java
@@ -14,7 +14,7 @@ public class OutsideRequest {
         this.direction = direction;
     }
 
-    public long getFloorNumber() {
+    public int getFloorNumber() {
         return floorNumber;
     }
